Cleanup sphinx and readthedocs configurations
* Use dirhtml as default builder for readthedocs (clean URLs)
* The configuration variables are now sorted with respect to the Sphinx
  documentation
* Remove useless comments in conf.py
* Add new comments for each section, following Sphinx documentation
  order
* The code corresponding to videos have been moved from config to a
  dedicated extension.
* Use proper HTML theme options
* Exclude all files starting with `.`, `_` (sphinx convention)
* Use OpenSearch
