+++ .readthedocs.yaml
@@ -10,7 +10,7 @@ submodules:
   recursive: true
 
 sphinx:
-  builder: html
+  builder: dirhtml
   configuration: conf.py
   fail_on_warning: false
